backend/requirements.txt
asgiref==3.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
Markdown==3.3.6
phonenumbers==8.12.43
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
